requirements.txt
typer==0.12.3
rich==13.7.1
pyyaml==6.0.2
feedparser==6.0.11
pytrends==4.9.2
python-slugify==8.0.4
pytz==2024.1